refactor: auto-split test_muscle_equilibrium.py
